Added menu items to nav
