+++ src/components/Book.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import BookForm from './BookForm';
+
+const Book = () => (
+  <div><BookForm /></div>
+);
+
+export default Book;
