+++ src/components/BookForm.js
@@ -1,5 +1,12 @@
 import React from 'react';
 
-const BookForm = () => <div>BookForm</div>;
+const BookForm = () => (
+  <div>
+    <form>
+      <input type="text" placeholder="Add Book" />
+      <button type="submit"> Add Book</button>
+    </form>
+  </div>
+);
 
 export default BookForm;
