--- src/components/Books.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Books = () => (
-  <div>Books</div>
-);
-
-export default Books;
